Move network types header file to platform abstractions
