+++ demos/network_manager/iot_network_manager_private.h
@@ -30,7 +30,7 @@
 #ifndef AWS_IOT_NETWORK_H_
 #define AWS_IOT_NETWORK_H_
 
-#include "iot_network_types.h"
+#include "types/iot_network_types.h"
 #include "aws_iot_network_config.h"
 
 #include "platform/iot_network.h"
