+++ libraries/abstractions/platform/include/types/iot_network_types.h
